Update staff command checking and fix unknown message bugs

+++ src/listeners/message.ts
@@ -60,7 +60,7 @@ client.on("message", async message => {
     if (/(https:\/\/)([a-z]{2,}\.)?crowdin\.com\/profile?\/?\S{1,}/gi.test(message.content) && message.channel.id === "569178590697095168") { //verify
         message.react("798339571531382874") //icon_working
         await crowdinVerify(message.member!, message.content.match(/(https:\/\/)([a-z]{2,}\.)?crowdin\.com\/profile\/\S{1,}/gi)?.[0], true)
-        await message.delete()
+        if (!message.deleted) await message.delete()
         message.channel.messages.fetch()
             .then(messages => {
                 const fiMessages = messages.filter(msgs => msgs.author === message.author);
